requirements.txt
arrow==0.14.2
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
Django==2.2.2
django-redis==4.10.0
djangorestframework==3.9.4
psycopg2==2.8.3
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
six==1.12.0
sqlparse==0.3.0
xlrd==1.2.0
uWSGI==2.0.18
